Functional web app With bars an area
Functional version of the Web Application. Candidate to be V0.2.0. It can read files by year,month and day. Year files only contain 12 months and the counts per month, similarily with month and day files, they only contain the count of each day in the month and by steps of 30min respectively